+++ selendroid-common/src/test/java/io/selendroid/SelendroidCapabilitiesTest.java
@@ -45,10 +45,9 @@ public class SelendroidCapabilitiesTest {
   public void testDefaultInitialize() {
     SelendroidCapabilities capa = new SelendroidCapabilities();
     Assert.assertEquals(null, capa.getEmulator());
-    Assert.assertEquals(4, capa.asMap().size());
+    Assert.assertEquals(3, capa.asMap().size());
     Assert.assertEquals("selendroid", capa.getAutomationName());
     Assert.assertEquals("selendroid", capa.getBrowserName());
-    Assert.assertEquals("en_US", capa.getLocale());
     Assert.assertEquals("android", capa.getPlatformName());
   }
 }